+++ src/components/form/tag-input/index.js
@@ -27,14 +27,14 @@ var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-co
 
 var _typographyConfig = _interopRequireDefault(require("../../../constants/typography-config"));
 
-var _crossIcon = require("../../../components/graphics/icons/cross-icon");
+var _crossIcon = require("../../graphics/icons/cross-icon");
 
 var _screenConfig = require("../../../constants/screen-config");
 
 var StyledWrapper = _styledComponents.default.div.withConfig({
   displayName: "tag-input__StyledWrapper",
   componentId: "sc-1iqglpj-0"
-})(["display:flex;align-items:center;font-size:", ";box-sizing:border-box;border-width:", ";border-style:solid;border-radius:0;width:100%;border-color:", ";padding:var(--tagInput-padding-vertical,", ") var(--tagInput-padding-horizontal,", ");.k-Form-TagList__list{width:100%;display:flex;flex-wrap:wrap;gap:", ";list-style:none;padding:0;margin:0;}.k-Form-TagList__item{padding:0;margin:0;order:1;max-width:100%;overflow:hidden;}.k-Form-TagList__inputItem{flex:1 0 auto;order:2;&:last-of-type .k-Form-TagList__input:empty::before{content:attr(aria-placeholder);}}.k-Form-TagList__input{width:100%;height:100%;min-width:", ";display:block;line-height:1;padding:", " 0 ", ";", ";cursor:text;&:empty::before{color:", ";content:' ';}:focus{outline-color:transparent;}}.k-Form-TagList__tagItem{display:flex;color:", ";background-color:", ";border-radius:", ";height:var(--tagInput-tag-height,", ");align-items:center;transition:color 0.2s ease,background-color 0.2s ease;cursor:default;&:hover{background-color:", ";}}.k-Form-TagList__tag{", ";padding:0 ", " 0.15em ", ";line-height:1;}.k-Form-TagList__button{", ";border:0;padding:0;background-color:transparent;color:inherit;padding:0 ", " 0 ", ";align-self:stretch;display:flex;align-items:center;cursor:pointer;&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}&:focus-within{outline:", " solid ", ";outline-offset:", ";}&.k-Form-TagList--disabled{cursor:not-allowed;.k-Form-TagList__tagItem{color:", ";background-color:", ";}.k-Form-TagList__tag,.k-Form-TagList__button{cursor:not-allowed;}}&.k-Form-TagList--tiny{--tagInput-padding-vertical:", ";--tagInput-tag-height:", ";min-height:", ";}&.k-Form-TagList--regular{--tagInput-padding-vertical:", ";min-height:", ";}&.k-Form-TagList--big{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--huge{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--giant{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--orion{&.k-Form-TagList--tiny,&.k-Form-TagList--regular{border-radius:", ";}&.k-Form-TagList--big,&.k-Form-TagList--huge,&.k-Form-TagList--giant{border-radius:", ";@media (min-width:", "px){--tagInput-padding-horizontal:", ";border-radius:", ";font-size:", ";}}}"], (0, _typography.stepToRem)(-1), (0, _typography.pxToRem)(2), _colorsConfig.default.line1, (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(8), (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(7), (0, _typography.pxToRem)(6), _typographyConfig.default.fontStyles.light, _colorsConfig.default.font2, _colorsConfig.default.text1, _colorsConfig.default.primary5, (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(30), _colorsConfig.default.primary4, _typographyConfig.default.fontStyles.regular, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(10), _typographyConfig.default.fontStyles.bold, (0, _typography.pxToRem)(12), (0, _typography.pxToRem)(10), _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(-2), _colorsConfig.default.primary4, _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2), _colorsConfig.default.font2, _colorsConfig.default.line1, (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(40), (0, _typography.pxToRem)(5), (0, _typography.pxToRem)(50), (0, _typography.pxToRem)(60), (0, _typography.pxToRem)(32), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(70), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(70), (0, _typography.pxToRem)(36), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(80), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(70), (0, _typography.pxToRem)(40), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(90), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(6), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(30), (0, _typography.pxToRem)(8), (0, _typography.stepToRem)(0));
+})(["display:flex;align-items:center;font-size:", ";box-sizing:border-box;border-width:", ";border-style:solid;border-radius:0;width:100%;border-color:", ";padding:var(--tagInput-padding-vertical,", ") var(--tagInput-padding-horizontal,", ");.k-Form-TagList__list{width:100%;display:flex;flex-wrap:wrap;gap:", ";list-style:none;padding:0;margin:0;}.k-Form-TagList__item{padding:0;margin:0;order:1;max-width:100%;overflow:hidden;}.k-Form-TagList__inputItem{flex:1 0 auto;order:2;&:last-of-type .k-Form-TagList__input:empty::before{content:attr(aria-placeholder);}}.k-Form-TagList__input{width:100%;height:100%;min-width:", ";display:block;line-height:1;padding:", " 0 ", ";", ";cursor:text;&:empty::before{color:", ";content:' ';}:focus{outline-color:transparent;}}.k-Form-TagList__tagItem{display:flex;color:", ";background-color:", ";border-radius:", ";height:var(--tagInput-tag-height,", ");align-items:center;transition:color 0.2s ease,background-color 0.2s ease;cursor:default;&:hover{background-color:", ";}&.k-Form-TagList__tagItem--disabled{color:", ";background-color:", ";.k-Form-TagList__tag,.k-Form-TagList__button{cursor:not-allowed;}}}.k-Form-TagList__tag{", ";padding:0 ", " 0.15em ", ";line-height:1;}.k-Form-TagList__button{", ";border:0;background-color:transparent;color:inherit;padding:0 ", " 0 ", ";align-self:stretch;display:flex;align-items:center;cursor:pointer;&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}&:focus-within{outline:", " solid ", ";outline-offset:", ";}&.k-Form-TagList--disabled{cursor:not-allowed;.k-Form-TagList__tagItem{color:", ";background-color:", ";}.k-Form-TagList__tag,.k-Form-TagList__button{cursor:not-allowed;}}&.k-Form-TagList--tiny{--tagInput-padding-vertical:", ";--tagInput-tag-height:", ";min-height:", ";}&.k-Form-TagList--regular{--tagInput-padding-vertical:", ";min-height:", ";}&.k-Form-TagList--big{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--huge{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--giant{min-height:", ";--tagInput-tag-height:", ";@media (min-width:", "px){min-height:", ";font-size:", ";}}&.k-Form-TagList--orion{&.k-Form-TagList--tiny,&.k-Form-TagList--regular{border-radius:", ";}&.k-Form-TagList--big,&.k-Form-TagList--huge,&.k-Form-TagList--giant{border-radius:", ";@media (min-width:", "px){--tagInput-padding-horizontal:", ";border-radius:", ";font-size:", ";}}}"], (0, _typography.stepToRem)(-1), (0, _typography.pxToRem)(2), _colorsConfig.default.line1, (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(8), (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(7), (0, _typography.pxToRem)(6), _typographyConfig.default.fontStyles.light, _colorsConfig.default.font2, _colorsConfig.default.text1, _colorsConfig.default.primary5, (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(30), _colorsConfig.default.primary4, _colorsConfig.default.font2, _colorsConfig.default.line1, _typographyConfig.default.fontStyles.regular, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(10), _typographyConfig.default.fontStyles.bold, (0, _typography.pxToRem)(12), (0, _typography.pxToRem)(10), _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(-2), _colorsConfig.default.primary4, _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2), _colorsConfig.default.font2, _colorsConfig.default.line1, (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(40), (0, _typography.pxToRem)(5), (0, _typography.pxToRem)(50), (0, _typography.pxToRem)(60), (0, _typography.pxToRem)(32), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(70), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(70), (0, _typography.pxToRem)(36), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(80), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(70), (0, _typography.pxToRem)(40), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(90), (0, _typography.stepToRem)(0), (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(6), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(30), (0, _typography.pxToRem)(8), (0, _typography.stepToRem)(0));
 
 var TagInput = function TagInput(_ref) {
   var placeholder = _ref.placeholder,
@@ -80,11 +80,13 @@ var TagInput = function TagInput(_ref) {
     });
   };
 
-  var removeValueFromList = function removeValueFromList(value) {
-    setLastRemoved(value);
+  var removeValueFromList = function removeValueFromList(item) {
+    var valueToRemove = (item === null || item === void 0 ? void 0 : item.value) || item;
+    setLastRemoved(valueToRemove);
     setItemList(function (currentList) {
       return currentList.filter(function (item) {
-        return item !== value;
+        var itemValue = (item === null || item === void 0 ? void 0 : item.value) || item;
+        return itemValue !== valueToRemove;
       });
     });
   };
@@ -144,21 +146,25 @@ var TagInput = function TagInput(_ref) {
     onKeyDown: onKeyDown,
     className: "k-Form-TagList__input"
   })), itemsList.map(function (item) {
+    var itemValue = (item === null || item === void 0 ? void 0 : item.value) || item;
+    var itemDisabled = (item === null || item === void 0 ? void 0 : item.disabled) || false;
     return /*#__PURE__*/_react.default.createElement("li", {
-      key: item,
-      className: "k-Form-TagList__item k-Form-TagList__tagItem"
+      key: itemValue,
+      className: (0, _classnames.default)('k-Form-TagList__item k-Form-TagList__tagItem', {
+        'k-Form-TagList__tagItem--disabled': itemDisabled
+      })
     }, /*#__PURE__*/_react.default.createElement("span", {
       className: "k-Form-TagList__tag"
-    }, item), /*#__PURE__*/_react.default.createElement("button", {
+    }, itemValue), /*#__PURE__*/_react.default.createElement("button", {
       className: "k-Form-TagList__button",
       type: "button",
-      disabled: disabled,
+      disabled: itemDisabled || disabled,
       onClick: function onClick() {
         return removeValueFromList(item);
       }
     }, /*#__PURE__*/_react.default.createElement("span", {
       className: "k-u-a11y-visuallyHidden"
-    }, "Retirer ", item, " de la liste."), /*#__PURE__*/_react.default.createElement(_crossIcon.CrossIcon, {
+    }, "Retirer ", itemValue, " de la liste."), /*#__PURE__*/_react.default.createElement(_crossIcon.CrossIcon, {
       color: "currentColor"
     })));
   })), /*#__PURE__*/_react.default.createElement("ul", {
@@ -167,9 +173,10 @@ var TagInput = function TagInput(_ref) {
     "aria-atomic": "true",
     "aria-relevant": "additions removals"
   }, itemsList.map(function (item) {
+    var itemValue = (item === null || item === void 0 ? void 0 : item.value) || item;
     return /*#__PURE__*/_react.default.createElement("li", {
-      key: "visuallyHidden-".concat(item)
-    }, item);
+      key: "visuallyHidden-".concat(itemValue)
+    }, itemValue);
   })));
 };
 
@@ -186,7 +193,7 @@ TagInput.defaultProps = {
 };
 TagInput.propTypes = {
   id: _propTypes.default.string.isRequired,
-  initialItemsList: _propTypes.default.arrayOf(_propTypes.default.string),
+  initialItemsList: _propTypes.default.arrayOf(_propTypes.default.any),
   addEventKeys: _propTypes.default.arrayOf(_propTypes.default.string),
   removeEventKeys: _propTypes.default.arrayOf(_propTypes.default.string),
   placeholder: _propTypes.default.string,