+++ src/esm/components/form/dropdown-select/combobox.js
@@ -53,7 +53,10 @@ export var DropdownCombobox = function DropdownCombobox(_ref) {
       className = _ref.className,
       value = _ref.value,
       _onBlur = _ref.onBlur,
-      controlled = _ref.controlled;
+      controlled = _ref.controlled,
+      modifier = _ref.modifier,
+      direction = _ref.direction,
+      arrowPosition = _ref.arrowPosition;
 
   var _useState = useState([]),
       _useState2 = _slicedToArray(_useState, 2),
@@ -164,7 +167,7 @@ export var DropdownCombobox = function DropdownCombobox(_ref) {
     setFlattenedOptions(flatOptions);
   }, [options]);
   return /*#__PURE__*/React.createElement(StyledDropdown, {
-    className: classNames('k-Form-Dropdown', "k-Form-Dropdown--".concat(variant), "k-Form-Dropdown--".concat(size), className, {
+    className: classNames('k-Form-Dropdown', "k-Form-Dropdown--".concat(variant), "k-Form-Dropdown--".concat(modifier), "k-Form-Dropdown--".concat(direction), "k-Form-Dropdown--".concat(size), "k-Form-Dropdown--arrowPosition-".concat(arrowPosition), className, {
       'k-Form-Dropdown--isOpen': isOpen > 0,
       'k-Form-Dropdown--error': error,
       'k-Form-Dropdown--valid': valid,
@@ -251,7 +254,10 @@ DropdownCombobox.defaultProps = {
   onMenuClose: function onMenuClose() {},
   onMenuOpen: function onMenuOpen() {},
   openOnLoad: false,
-  menuZIndex: 1000
+  menuZIndex: 1000,
+  modifier: 'hydrogen',
+  direction: 'down',
+  arrowPosition: 'left'
 };
 DropdownCombobox.propTypes = {
   id: PropTypes.string.isRequired,
@@ -264,7 +270,7 @@ DropdownCombobox.propTypes = {
   placeholder: PropTypes.string,
   labelPropsGetter: PropTypes.func,
   variant: PropTypes.oneOf(['andromeda', 'orion']),
-  size: PropTypes.oneOf(['tiny', 'normal', 'big', 'huge', 'giant']),
+  size: PropTypes.oneOf(['micro', 'tiny', 'normal', 'big', 'huge', 'giant']),
   a11yStatusError: PropTypes.string,
   a11yStatusValid: PropTypes.string,
   a11ySelectionMessageDisplayer: PropTypes.func,
@@ -273,5 +279,8 @@ DropdownCombobox.propTypes = {
   onMenuClose: PropTypes.func,
   onMenuOpen: PropTypes.func,
   openOnLoad: PropTypes.bool,
-  menuZIndex: PropTypes.number
+  menuZIndex: PropTypes.number,
+  modifier: PropTypes.oneOf(['hydrogen', 'nitrogen', 'boron']),
+  direction: PropTypes.oneOf(['up', 'down']),
+  arrowPosition: PropTypes.oneOf(['left', 'right'])
 };